Move ViewManagerModel to the general interface_adapter package
